+++ dynamic_accounts_report/static/src/js/trial_balance.js
@@ -9,6 +9,9 @@ odoo.define('dynamic_cash_flow_statements.trial', function (require) {
     var QWeb = core.qweb;
     var _t = core._t;
 
+        var datepicker = require('web.datepicker');
+    var time = require('web.time');
+
     window.click_num = 0;
     var TrialBalance = AbstractAction.extend({
     template: 'TrialTemp',
@@ -19,6 +22,7 @@ odoo.define('dynamic_cash_flow_statements.trial', function (require) {
             'click #pdf': 'print_pdf',
             'click #xlsx': 'print_xlsx',
             'click .show-gl': 'show_gl',
+            'mousedown div.input-group.date[data-target-input="nearest"]': '_onCalendarIconClick',
         },
 
         init: function(parent, action) {
@@ -43,6 +47,36 @@ odoo.define('dynamic_cash_flow_statements.trial', function (require) {
                 self.load_data(self.initial_render);
             })
         },
+
+        _onCalendarIconClick: function (ev) {
+
+        var $calendarInputGroup = $(ev.currentTarget);
+
+        var calendarOptions = {
+
+//        minDate: moment({ y: 1000 }),
+//            maxDate: moment().add(200, 'y'),
+//            calendarWeeks: true,
+//            defaultDate: moment().format(),
+//            sideBySide: true,
+//            buttons: {
+//                showClear: true,
+//                showClose: true,
+//                showToday: true,
+//            },
+
+            icons : {
+                date: 'fa fa-calendar',
+
+            },
+            locale : moment.locale(),
+            format : time.getLangDateFormat(),
+             widgetParent: 'body',
+             allowInputToggle: true,
+        };
+
+        $calendarInputGroup.datetimepicker(calendarOptions);
+    },
 
 
         load_data: function (initial_render = true) {
@@ -223,13 +257,21 @@ odoo.define('dynamic_cash_flow_statements.trial', function (require) {
             }
             filter_data_selected.journal_ids = journal_ids
 
-            if ($("#date_from").val()) {
-                var dateString = $("#date_from").val();
-                filter_data_selected.date_from = dateString;
+//            if ($("#date_from").val()) {
+//                var dateString = $("#date_from").val();
+//                filter_data_selected.date_from = dateString;
+//            }
+//            if ($("#date_to").val()) {
+//                var dateString = $("#date_to").val();
+//                filter_data_selected.date_to = dateString;
+//            }
+
+             if (this.$el.find('.datetimepicker-input[name="date_from"]').val()) {
+                filter_data_selected.date_from = moment(this.$el.find('.datetimepicker-input[name="date_from"]').val(), time.getLangDateFormat()).locale('en').format('YYYY-MM-DD');
             }
-            if ($("#date_to").val()) {
-                var dateString = $("#date_to").val();
-                filter_data_selected.date_to = dateString;
+
+            if (this.$el.find('.datetimepicker-input[name="date_to"]').val()) {
+                filter_data_selected.date_to = moment(this.$el.find('.datetimepicker-input[name="date_to"]').val(), time.getLangDateFormat()).locale('en').format('YYYY-MM-DD');
             }
 
             if ($(".target_move").length) {