+++ dynamic_accounts_report/wizard/balance_sheet.py
@@ -38,6 +38,7 @@ class BalanceSheetView(models.TransientModel):
 
     @api.model
     def view_report(self, option, tag):
+
         r = self.env['dynamic.balance.sheet.report'].search(
             [('id', '=', option[0])])
         data = {
@@ -83,9 +84,15 @@ class BalanceSheetView(models.TransientModel):
 
             new_records = list(filter(filter_code, records['Accounts']))
             records['Accounts'] = new_records
+        trans_tag = self.env['ir.translation'].search([('value', '=', tag), ('module', '=', 'dynamic_accounts_report')],
+                                                      limit=1).src
+        if trans_tag:
+            tag_upd = trans_tag
+        else:
+            tag_upd = tag
 
-        account_report_id = self.env['account.financial.report'].search([
-            ('name', 'ilike', tag)])
+        account_report_id = self.env['account.financial.report'].with_context(lang='en_US').search([
+            ('name', 'ilike', tag_upd)])
 
         new_data = {'id': self.id, 'date_from': False,
                     'enable_filter': True,
@@ -161,7 +168,6 @@ class BalanceSheetView(models.TransientModel):
 
         parent_list = list(set(parent_list))
         final_report_lines = []
-
         for rec in report_lines_move:
             if rec['report_type'] != 'accounts':
                 if rec['r_id'] in parent_list:
@@ -220,7 +226,6 @@ class BalanceSheetView(models.TransientModel):
                 rec['m_credit'] = "{:,.2f}".format(rec['credit']) + " " + symbol
                 rec['m_balance'] = "{:,.2f}".format(
                     rec['balance']) + " " + symbol
-
         return {
             'name': tag,
             'type': 'ir.actions.client',