+++ dynamic_accounts_report/doc/RELEASE_NOTES.md
@@ -11,5 +11,15 @@
 #### UPDT
 - Updated
 
+#### 20.12.2021
+#### Version 13.0.1.0.2
+#### UPDT
+- Translation issue and Calendar format issue
+
+#### 15.01.2022
+#### Version 13.0.1.0.3
+#### UPDT
+- Arabic Translation added
+